// backend/src/model/ECPay.ts
export type ECPayOrderResult = {
  MerchantID: string;
  MerchantTradeNo: string;
  PaymentDate: string;
  PaymentType: string;
  PaymentTypeChargeFee: string;
  RtnCode: string;
  RtnMsg: string;
  SimulatePaid: string;
  StoreId: string;
  TradeAmt: string;
  TradeDate: string;
  TradeNo: string;
  CheckMacValue: string;
  CustomField1: string;
  CustomField2: string;
  CustomField3: string;
  CustomField4: string;
};